fix: Fade deleted images fix: Don't include empty paragraphs as context fix: Allow for same image multiple times and refactor
